* commit '76cef14091e28264f00c317aba2e2d0bbc86843a': (65 commits)
  CAR-1074 SQL fix
  CAR-1074 Content change for Virgin
  CAR-1076 Enable sessioncam
  CAR-1065 Fix .end()
  CAR-1075: Modified to better handle hiding when real/wool and whether excess is array or object
  HLT-2453:
  CAR-1068: modified to force rego to be uppercased in java
  TRV-875: Fixed some pricing based on user preview feedback
  UTL-297: Add extra things
  CAR-1067: Modifed AI's online only products to remove the phone number from response
  TRV-875: Modified product benefits to use the providerProductCode rather than the productId
  UTL-297: Change effective end
  CAR-1068 - Enable MotorWeb in all environments
  CAR-1068 - Remove unused import reference
  CAR-1068 - MotorWeb certs
  CAR-1068 - Updates to Service config
  CAR-1067: Omitting the telephone number from response when ONLINE ONLY
  HLT-2332: fix coupon optin gets reset when change steps
  CAR-1068 - Update method for loading MotorWeb Certificate to allow to load from external file system if required.
  HNC-290: Made fact sheet paths relative to accommodate el-Choosi
  ...

+++ src/com/ctm/services/simples/SimplesSearchService.java
@@ -128,8 +128,8 @@ public class SimplesSearchService {
         String sql;
         SqlDao<Object> sqlDao = new SqlDao<>();
         DatabaseQueryMapping<Object> mapping;
-        final List<Integer> transactionIDsHot = new ArrayList<>();
-        final List<Integer> transactionIDsCold = new ArrayList<>();
+        final List<Long> transactionIDsHot = new ArrayList<>();
+        final List<Long> transactionIDsCold = new ArrayList<>();
         sql = "SELECT rootId AS id\n" +
                 "\t\t\t\t\tFROM aggregator.transaction_header\n" +
                 "\t\t\t\t\tWHERE  TransactionID = ?\n" +
@@ -148,10 +148,10 @@ public class SimplesSearchService {
 
             @Override
             public Object handleResult(ResultSet rs) throws SQLException {
-                return rs.getInt(1);
+                return rs.getLong(1);
             }
         };
-        final int rootID = (int) sqlDao.get(mapping, sql);
+        final Long rootID = (Long) sqlDao.get(mapping, sql);
         if (rootID != 0) {
             sql = "SELECT 'HOT' as tableType , transactionId AS id\n" +
                     "\t\t\t\t\t\tFROM aggregator.transaction_header\n" +
@@ -171,9 +171,9 @@ public class SimplesSearchService {
                 public Object handleResult(ResultSet rs) throws SQLException {
                     while (rs.next()) {
                         if (rs.getString("tableType") != null && rs.getString("tableType").equalsIgnoreCase("hot"))
-                            transactionIDsHot.add(rs.getInt("id"));
+                            transactionIDsHot.add(rs.getLong("id"));
                         else
-                            transactionIDsCold.add(rs.getInt("id"));
+                            transactionIDsCold.add(rs.getLong("id"));
                     }
                     return null;
                 }
@@ -192,7 +192,7 @@ public class SimplesSearchService {
         String sql;
         SqlDao<Object> sqlDao = new SqlDao<>();
         DatabaseQueryMapping<Object> mapping;
-        List<Integer> transactionIDs;
+        List<Long> transactionIDs;
         sql = "SELECT distinct transactionId as id\n" +
                 "\t\t\t\t\tFROM aggregator.phone_lookup\n" +
                 "\t\t\t\t\twhere PhoneNumber =  ?\n" +
@@ -205,14 +205,14 @@ public class SimplesSearchService {
 
             @Override
             public Object handleResult(ResultSet rs) throws SQLException {
-                List<Integer> transactionIDs = new ArrayList<>();
+                List<Long> transactionIDs = new ArrayList<>();
                 while (rs.next()) {
-                    transactionIDs.add(rs.getInt(1));
+                    transactionIDs.add(rs.getLong(1));
                 }
                 return transactionIDs;
             }
         };
-        transactionIDs = (List<Integer>) sqlDao.getAll(mapping, sql);
+        transactionIDs = (List<Long>) sqlDao.getAll(mapping, sql);
         hotTransactionIdsCsv = StringUtils.join(transactionIDs, ",");
     }
 
@@ -224,7 +224,7 @@ public class SimplesSearchService {
         String sql;
         SqlDao<Object> sqlDao = new SqlDao<>();
         DatabaseQueryMapping<Object> mapping;
-        List<Integer> transactionIDs;
+        List<Long> transactionIDs;
         sql = "\tSELECT distinct transactionId as id\n" +
                 "\t\t\t\t\tFROM aggregator.email_lookup\n" +
                 "\t\t\t\t\twhere emailAddress =  ?\n" +
@@ -237,14 +237,14 @@ public class SimplesSearchService {
 
             @Override
             public Object handleResult(ResultSet rs) throws SQLException {
-                List<Integer> transactionIDs = new ArrayList<>();
+                List<Long> transactionIDs = new ArrayList<>();
                 while (rs.next()) {
-                    transactionIDs.add(rs.getInt("id"));
+                    transactionIDs.add(rs.getLong("id"));
                 }
                 return transactionIDs;
             }
         };
-        transactionIDs = (List<Integer>) sqlDao.getAll(mapping, sql);
+        transactionIDs = (List<Long>) sqlDao.getAll(mapping, sql);
         hotTransactionIdsCsv = StringUtils.join(transactionIDs, ",");
     }
 
@@ -256,7 +256,7 @@ public class SimplesSearchService {
         String sql;
         SqlDao<Object> sqlDao = new SqlDao<>();
         DatabaseQueryMapping<Object> mapping;
-        List<Integer> transactionIDs;
+        List<Long> transactionIDs;
         if (searchString.split(" ").length > 1) {
             final String firstName = searchString.split(" ")[0];
             final String lastName = searchString.split(" ")[1];
@@ -280,14 +280,14 @@ public class SimplesSearchService {
 
                 @Override
                 public Object handleResult(ResultSet rs) throws SQLException {
-                    List<Integer> transactionIDs = new ArrayList<>();
+                    List<Long> transactionIDs = new ArrayList<>();
                     while (rs.next()) {
-                        transactionIDs.add(rs.getInt(1));
+                        transactionIDs.add(rs.getLong(1));
                     }
                     return transactionIDs;
                 }
             };
-            transactionIDs = (List<Integer>) sqlDao.getAll(mapping, sql);
+            transactionIDs = (List<Long>) sqlDao.getAll(mapping, sql);
         } else {
 
             sql = "SELECT distinct transactionId as id\n" +
@@ -303,14 +303,14 @@ public class SimplesSearchService {
 
                 @Override
                 public Object handleResult(ResultSet rs) throws SQLException {
-                    List<Integer> transactionIDs = new ArrayList<>();
+                    List<Long> transactionIDs = new ArrayList<>();
                     while (rs.next()) {
-                        transactionIDs.add(rs.getInt(1));
+                        transactionIDs.add(rs.getLong(1));
                     }
                     return transactionIDs;
                 }
             };
-            transactionIDs = (List<Integer>) sqlDao.getAll(mapping, sql);
+            transactionIDs = (List<Long>) sqlDao.getAll(mapping, sql);
         }
         hotTransactionIdsCsv = StringUtils.join(transactionIDs, ",");
     }
